requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
google-generativeai
openai

google-cloud-aiplatform
protobuf